Renamed original Signer to SignatureAlgorithm for disambiguation
