+++ src/jose/algorithms/signing/signature-algorithms.ts
@@ -1,10 +1,10 @@
-import type { Signer } from '../../../types/jose-types.js';
+import type { SignatureAlgorithm } from '../../../types/jose-types.js';
 
 import { ed25519 } from './ed25519.js';
 import { Secp256k1 } from '../../../utils/secp256k1.js';
 
 // the key should be the appropriate `crv` value
-export const signers: Record<string, Signer> = {
+export const signatureAlgorithms: Record<string, SignatureAlgorithm> = {
   'Ed25519'   : ed25519,
   'secp256k1' : {
     sign            : Secp256k1.sign,
