+++ src/utils/private-key-signer.ts
@@ -1,7 +1,7 @@
 import type { PrivateJwk } from '../types/jose-types.js';
 import type { Signer } from '../types/signer.js';
 
-import { signers as signatureAlgorithms } from '../jose/algorithms/signing/signers.js';
+import { signatureAlgorithms } from '../jose/algorithms/signing/signature-algorithms.js';
 import { DwnError, DwnErrorCode } from '../index.js';
 
 /**
